pyproject.toml
###########
# 📜 Poetry
###########
[tool.poetry]
name = "jizhu"
version = "0.1.0"
description = "用 Anki 记住中文单词"
authors = ["John Sandall <contact@coefficient.ai>"]
license = "UNLICENSED"
classifiers = ["Private :: Do Not Upload"]
packages = [
  { include = "jizhu", from = "jizhu" },
  # Add more packages here, e.g.
  # { include = "module_name", from = "jizhu" },
]

[tool.poetry.dependencies]
python = ">=3.11.2,<3.12"
# Everything below here is alphabetically sorted
genanki = "^0.13.0"
google-cloud-translate = "^3.12.0"
ipykernel = "^6.25.1"
joblib = "^1.3.2"
jupyterlab = "^4.0.5"
loguru = "^0.7.0"
numpy = "^1.25.2"
openpyxl = "^3.1.2"
pandas = "^2.0.3"
pypinyin = "^0.49.0"
scipy = "^1.11.2"
typer = "^0.9.0"

[tool.poetry.dev-dependencies]
# Everything below here is alphabetically sorted
add-trailing-comma = "^3.0.1"
bandit = "^1.7.5"
black = "^23.7.0"
detect-secrets = "1.2.0"
flake8 = "^6.1.0"
flake8-bugbear = "^23.7.10"
flake8-comprehensions = "^3.14.0"
flake8-docstrings = "^1.7.0"
flake8-eradicate = "^1.5.0"
flake8-fixme = "^1.1.1"
flake8-implicit-str-concat = "^0.4.0"
flake8-mutable = "^1.2.0"
flake8-no-pep420 = "^2.7.0"
flake8-print = "^5.0.0"
flake8-return = "^1.2.0"
ipdb = "^0.13.13"
isort = "^5.12.0"
jupyter-black = "^0.3.4"
mypy = "^1.5.1"
pip-audit = "^2.6.1"
pre-commit = "^3.3.3"
pylint = "^2.17.5"
pytest = "^7.4.0"
towncrier = "^23.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

############
# ✅ Linters
############
[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100
default_section = "THIRDPARTY"
known_first_party = ["jizhu"]
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
add_imports = ""

[tool.pylint.MASTER]
ignore-paths = [".git"]
load-plugins = []

[tool.pylint."MESSAGES CONTROL"]
enable = "all"
max-module-lines = 2000
max-line-length = 100
max-locals = 50
min-similarity-lines = 150
max-statements = 89
max-args = 22
max-branches = 17
good-names = ['df', 'i', 'py']
disable = ['missing-function-docstring']
logging-format-style = "new"

##############
# 📣 Towncrier
##############
[tool.towncrier]
package = "jizhu"
package_dir = "."
filename = "CHANGELOG.md"
directory = ".changelog/"
template = ".changelog/template.md"
title_format = "## [{version}] - {project_date}"
issue_format = "[#{issue}](https://github.com/CoefficientSystems/c459-jizhu/issues/{issue})"
start_string = "<!-- TOWNCRIER -->\n"
underlines = ["", ""]

# .added for new features
# .changed for changes in existing functionality
# .deprecated for soon-to-be removed features
# .removed for now removed features
# .fixed for any bug fixes
# .security in case of vulnerabilities
# .analysis for data analyses
# .docs for documentation improvements
# .maintenance for maintenance tasks & upgrades
[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "deprecated"
name = "Deprecated"
showcontent = true

[[tool.towncrier.type]]
directory = "removed"
name = "Removed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Bug Fixes"
showcontent = true

[[tool.towncrier.type]]
directory = "security"
name = "Security"
showcontent = true

[[tool.towncrier.type]]
directory = "analysis"
name = "Analyses"
showcontent = true

[[tool.towncrier.type]]
directory = "docs"
name = "Improved Documentation"
showcontent = true

[[tool.towncrier.type]]
directory = "maintenance"
name = "Maintenance Changes"
showcontent = true
